refactor: teachingbubble
